requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
sqlalchemy==2.0.37
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
reportlab==4.3.0
aiofiles==24.1.0
pydantic[email]>=2.0
